Refactoring - code cleanup, method ordering, renaming files and classes
